Reorganize tests in test_session module
